encoder: h264: Export SlicePromise as BitstreamPromise
